requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Deprecated==1.2.13
dj-database-url==1.3.0
Django==4.2
django-appconf==1.0.5
django-cors-headers==3.14.0
django-filter==23.1
django-imagekit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pilkit==2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vonage==3.3.0
whitenoise==6.2.0
wrapt==1.15.0
